[roborock] Guard map dimensions against integer overflow
The image width and height are unvalidated uint32 values read straight
from the map payload, but both length guards multiplied them in int
arithmetic. A header claiming 65536 x 65536 wraps to exactly 0, one
claiming 65536 x 65535 wraps to a negative number, and either passes a
guard of the form "imageData.length < width * height".

In RoomAtRobotResolver that matters twice over, because the pixel lookup
behind the guard only compares against width and height and never
against the array length. Depending on the declared dimensions and the
size of the real image data, the lookup then either throws
ArrayIndexOutOfBoundsException - which would travel up into the
transport callback, where only JsonSyntaxException is caught - or, worse,
reads a pixel far outside the intended row and publishes whatever
segment happens to live there as the robot's room. Nothing about that
second outcome looks like a failure.

Multiplying as a long fixes both: a payload with a lying header is
rejected, and a guard that passes now implies width * height <=
imageData.length <= Integer.MAX_VALUE, so every in-bounds pixel index
fits into an int as well.

RRMapRenderer carries the same guard - it is where the resolver's copy
came from - and gets the same fix. There a wrapped product reached the
image allocation behind it instead.

AI-assisted-by: Claude Code

+++ bundles/org.openhab.binding.roborock/src/main/java/org/openhab/binding/roborock/internal/map/RRMapRenderer.java
@@ -81,7 +81,10 @@ public final class RRMapRenderer {
         int width = mapData.imageWidth();
         int height = mapData.imageHeight();
         byte[] imageData = mapData.imageData();
-        if (width <= 0 || height <= 0 || imageData.length < width * height) {
+        // Multiplied as a long: the dimensions are unvalidated uint32 values from the map payload,
+        // and a product that wraps in int arithmetic would let a payload with a lying header past
+        // this guard.
+        if (width <= 0 || height <= 0 || imageData.length < (long) width * height) {
             throw new RoborockException("Cannot render map image due to invalid dimensions or data length.");
         }
